- Updated calling standard doc - Functional perror() - Functional open() and close() for unbuffered channels - DEBUG conditional compilation variable, KDPRINT, KDPUTSTR

+++ lib/sys/lib_getpid.c
@@ -1,9 +1,10 @@
 #include "muxx.h"
 #include "muxxlib.h"
 
-WORD getpid() {
+int getpid() {
   int rc;
   TCB thisTask;
+
   if ((rc = gettpi(0,&thisTask)) == 0) {
     return thisTask.pid;
   }
